start moving view
