+++ crates/nix_rs/src/flake/mod.rs
@@ -6,8 +6,6 @@ pub mod show;
 pub mod system;
 pub mod url;
 
-use leptos::*;
-use leptos_router::*;
 use serde::{Deserialize, Serialize};
 
 use self::{outputs::FlakeOutputs, schema::FlakeSchema, system::System, url::FlakeUrl};
@@ -23,21 +21,4 @@ pub struct Flake {
     /// Flake output schema (typed version of [FlakeOutputs])
     pub schema: FlakeSchema,
     // TODO: Add `nix flake metadata` info.
-}
-
-impl IntoView for Flake {
-    fn into_view(self, cx: Scope) -> View {
-        view! { cx,
-            <div class="flex flex-col my-4">
-                <h3 class="text-lg font-bold">{self.url}</h3>
-                <div class="text-sm italic text-gray-600">
-                    <A href="/flake/raw" exact=true>
-                        "View raw output"
-                    </A>
-                </div>
-                <div>{self.schema}</div>
-            </div>
-        }
-        .into_view(cx)
-    }
 }
